pyproject.toml
[project]
name = "blender-mcp"
version = "1.2.1"
description = "Model Context Protocol server for Blender 3D integration"
readme = "README.md"
authors = [
    { name = "BlenderMCP Contributors", email = "contact@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["mcp", "blender", "3d", "ai", "llm", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
]

dependencies = [
    "mcp[cli]>=1.0.0,<2.0.0",
    "fake-bpy-module>=20230101",  # For type hints
    "requests>=2.25.0",
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.6.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "blender-mcp[gui,test]",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
]

[project.scripts]
blender-mcp = "blender_mcp.cli:main"
blender-mcp-gui = "blender_mcp.gui:main"

[project.urls]
Homepage = "https://github.com/modelcontextprotocol/blender-mcp"
Repository = "https://github.com/modelcontextprotocol/blender-mcp"
Issues = "https://github.com/modelcontextprotocol/blender-mcp/issues"
Documentation = "https://github.com/modelcontextprotocol/blender-mcp#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/blender_mcp", "addon"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "e2e: end-to-end tests that use real external processes (e.g., Blender)",
    "visual: visual regression tests based on screenshot baseline",
]

[tool.coverage.run]
source = ["src/blender_mcp"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__
"tests/*" = ["F401", "F811"]  # Allow test fixtures

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
